requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
Django==3.2.18
django-crispy-forms==2.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
